trytond : Add update parameter to post init hooks
Some hooks will not be necessary on updates, others will only be required when
updating some modules, etc... So we add an 'update' argument to hooks so that
it becomes possible to activate the hook or not depending on the context.

Consequently, we move the logging part from the post_init method to inside the
hook, so the log is not filled with hooks which aren't actually triggered.