Do not set _timestamp on fake singleton instance
It is not really useful because the id of the future record can not be guessed
for sure by the client.
Also when write is called with such fake _timestamp as singleton create the
record before actually write, the timestamp check always fails (if the id was
correctly guessed).

issue10282
(grafted from cea2a39c6e31f3aa2405cf40d39951c52e1f9be5)